requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
Pillow==10.2.0
python-dotenv==1.0.1
django-allauth==0.61.1
django-cleanup==8.1.0
whitenoise==6.6.0
gunicorn==21.2.0 
# newss
feedparser==6.0.10
beautifulsoup4
django-tinymce

# 
scikit-learn

